% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genFakeData.R
\name{genFakeData}
\alias{genFakeData}
\title{genFakeData}
\usage{
genFakeData(fake_data, mu_mean = 700, mu_sigma = 75, gamma_mean = 0.5,
  gamma_sigma = 1, sigma_mean = 16, sigma_sigma = 11,
  sigma_a_sigma = 100)
}
\arguments{
\item{fake_data}{}

\item{mu_mean}{}

\item{mu_sigma}{}

\item{gamma_mean}{}

\item{gamma_sigma}{}

\item{sigma_mean}{}

\item{sigma_sigma}{}

\item{sigma_a_sigma}{}
}
\description{
genFakeData
}
